+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -68,6 +68,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Changed Receive SSR to also controls Receive SSGI on opaque objects.
 - Improved the punctual light shadow rescale algorithm.
 - Changed the names of some of the parameters for the Eye Utils SG Nodes.
+- Restored frame setting for async compute of contact shadows.
 
 ## [10.1.0] - 2020-10-12